+++ src/distVector.cpp
@@ -1199,7 +1199,7 @@ SpatGeom hullify(SpatVector b, bool ispoly) {
 	SpatVector part;
 	part.reserve(b.size());
 	for (size_t j =0; j<(b.size()-1); j++) {
-		std::vector<unsigned> range = {(unsigned)j, (unsigned)j+1};
+		std::vector<size_t> range = {j, j+1};
 		SpatVector g = b.subset_rows(range);
 		g = g.hull("convex");
 		part.addGeom(g.geoms[0]);
@@ -1283,7 +1283,7 @@ SpatVector lonlat_buf(SpatVector x, double dist, unsigned quadsegs, bool ispol, 
 SpatVector SpatVector::buffer_lonlat(std::string vt, std::vector<double> d, unsigned quadsegs) {
 
 	SpatVector out;
-	std::vector<unsigned> keep;
+	std::vector<size_t> keep;
 	keep.reserve(size());
 	if (vt == "points") {
 		return point_buffer(d, quadsegs, false, true);
