+++ src/table_utils.h
@@ -1,6 +1,6 @@
 
 
-std::map<double, unsigned long long> table(std::vector<double> &v);
-std::map<double, unsigned long long int> combine_tables(std::map<double, unsigned long long int> &x, std::map<double, unsigned long long int> &y);
-std::vector<double> table2vector(std::map<double, unsigned long long int> &x);
-std::vector<std::vector<double>> table2vector2(std::map<double, unsigned long long int> &x);
+std::map<double, size_t> table(std::vector<double> &v);
+std::map<double, size_t> combine_tables(std::map<double, size_t> &x, std::map<double, size_t> &y);
+std::vector<double> table2vector(std::map<double, size_t> &x);
+std::vector<std::vector<double>> table2vector2(std::map<double, size_t> &x);
